+++ vedo/core.py
@@ -428,7 +428,7 @@ class CommonAlgorithms:
         return self.dataset.GetActualMemorySize()
 
     def modified(self):
-        """Use in conjunction with ``tonumpy()`` to update any modifications to the image array"""
+        """Use in conjunction with `tonumpy()` to update any modifications to the image array"""
         self.dataset.GetPointData().Modified()
         self.dataset.GetPointData().GetScalars().Modified()
         return self